feat: Implement config loader (#143)
* feat : check ports in a Windows environment

* chore : refactor build_config

* feat: Implement config loader

---------
