+++ examples/memstore/common_raftnode_config.toml
@@ -1,0 +1,9 @@
+log_dir = "./logs"
+compacted_log_size_threshold = 1073741824
+save_compacted_logs = true
+
+[raft_config]
+tick_interval = 0.2
+election_tick = 10
+heartbeat_tick = 3
+omit_heartbeat_log = false
