+++ examples/src/config.rs
@@ -1,4 +1,6 @@
-use raftify::{Config, ConfigBuilder, Peers, RaftConfig};
+use std::path::Path;
+
+use raftify::{load_configs, Config, ConfigBuilder, Peers};
 
 #[cfg(feature = "tls")]
 use raftify::TlsConfig;
@@ -6,26 +8,22 @@ use raftify::TlsConfig;
 use crate::utils::{ensure_directory_exist, get_storage_path};
 
 pub fn build_config(node_id: u64, initial_peers: Option<Peers>) -> Config {
-    let raft_config = RaftConfig {
-        id: node_id,
-        election_tick: 10,
-        heartbeat_tick: 3,
-        omit_heartbeat_log: false,
-        ..Default::default()
-    };
-
     let storage_pth = get_storage_path("./logs", node_id);
     ensure_directory_exist(&storage_pth).expect("Failed to create storage directory");
 
-    #[allow(unused_mut)]
-    let mut config_builder = ConfigBuilder::new()
-        .log_dir("./logs".to_owned())
-        .save_compacted_logs(true)
-        .compacted_log_dir("./logs".to_owned())
-        .compacted_log_size_threshold(1024 * 1024 * 1024)
-        .raft_config(raft_config)
-        .tick_interval(0.2);
+    let path = Path::new(file!())
+        .parent()
+        .unwrap()
+        .parent()
+        .unwrap()
+        .join("memstore/common_raftnode_config.toml");
 
+    let config_builder = ConfigBuilder::from_config(
+        load_configs(path.to_str().unwrap()).expect("Failed to load common config"),
+    )
+    .set_node_id(node_id);
+
+    #[allow(unused_mut)]
     #[cfg(feature = "tls")]
     {
         let client_tls_config = TlsConfig {
