+++ raftify/Cargo.toml
@@ -16,7 +16,7 @@ bytes = "1.7.2"
 log = { version = "0.4", features = ["std"] }
 parking_lot = "0.12.3"
 prost = "0.11"
-raft = { version = "0.7.10", features = ["prost-codec", "default-logger"], default-features = false, package = "jopemachine-raft" }
+raft = { version = "0.7.13", features = ["prost-codec", "default-logger"], default-features = false, package = "jopemachine-raft" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 slog = "2"
@@ -29,6 +29,7 @@ chrono = "0.4.38"
 heed = { version = "0.20.5", optional = true }
 heed-traits = { version = "0.20", optional = true }
 rocksdb = { version = "0.19.0", optional = true }
+config = "0.14.0"
 
 [features]
 default = ["heed_storage", "tls"]
